requirements.txt
aniso8601==7.0.0
astroid==2.3.3
autopep8==1.5.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-GraphQL==2.0.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
grpcio==1.30.0
grpcio-tools==1.30.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
pathtools==0.1.2
promise==2.3
protobuf==3.12.2
pycodestyle==2.6.0
pycparser==2.20
pylint==2.4.4
pymongo==3.11.0
python-dotenv==0.13.0
Rx==1.6.1
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.3
